Bot_help/requirements.txt
aiogram
python-dotenv
openai
apscheduler
httpx
yt_dlp
groq
gigachat
beautifulsoup4
pypdf
SpeechRecognition
pydub
